refactor RocketFragment

+++ app/src/main/java/com/x5bartsoft/spacex/model/response/launchdetail/SecondStage.kt
@@ -12,7 +12,7 @@ data class SecondStage(
     @SerializedName("engines")
     val engines: Int,
     @SerializedName("fuel_amount_tons")
-    val fuelAmountTons: Int,
+    val fuelAmountTons: Double,
     @SerializedName("payloads")
     val payloads: Payloads,
     @SerializedName("reusable")
